requirements.txt
fastapi==0.115.0
pydantic==2.9.2
uvicorn[standard]==0.30.0
vellum-ai>=0.1.0
reportlab>=4.0.0
python-multipart>=0.0.6
